Changed costumer selection logic to App.js and its state. Rendering SelecitionBar on app.js too now.
