+++ src/components/SelectedCustomer.js
@@ -5,7 +5,6 @@ import PropTypes from 'prop-types';
 class SelectedCustomer extends Component {
 
   render () {
-
     return (
       <h4 > Selected Customer: {this.props.customerName} </h4>
     )
@@ -15,7 +14,7 @@ class SelectedCustomer extends Component {
 export default SelectedCustomer;
 
 SelectedCustomer.propTypes = {
-
+customerName:  PropTypes.string.isRequired,
 }
 
 // hidden class on css in a conditional statement:
